Add users and symptoms tables
